ui element position as a % and ui element scaling

+++ include/components.hpp
@@ -26,7 +26,7 @@ namespace Cascade
 
   struct UIElement
   {
-    int position[2]{0, 0};
+    float position[2]{0, 0};
     int size[2]{0, 0};
     float angle{0};
     bool click_type[3]{false, false, false}; // left, middle, right
